pyproject.toml
[project]
name = "pyside"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "evarle", email = ""},
]
dependencies = ["PySide6>=6.9.1", "Pillow>=11.3.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
